chore: refactor worker to es module
